NF: Introduction of different storage layouts

+++ platform.sh
@@ -4,6 +4,8 @@ BASEDIR=$(dirname "$0")
 
 KUBEMASTER_IPADDR="10.88.20.109"
 
+STORAGELAYOUT=`$BASEDIR/python/read-values.py`
+
 function attach_arbiterdisk () {
   VM=kubenode$1
   DEVICE=$2
@@ -22,7 +24,7 @@ function attach_arbiterdisk () {
   virsh attach-disk $VM --source $FILE --target $DEVICE --persistent --subdriver qcow2
 
   # Create a physical volume for the newly attached disk
-  ansible kubenode$1 -i $BASEDIR/ansible/inventory.yaml -a "pvcreate /dev/$DEVICE"
+  ansible kubenode$1 -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml -a "pvcreate /dev/$DEVICE"
 }
 
 function attach_datadisk () {
@@ -43,7 +45,7 @@ function attach_datadisk () {
   virsh attach-disk $VM --source $FILE --target $DEVICE --persistent --subdriver qcow2
 
   # Create a physical volume for the newly attached disk
-  ansible kubenode$1 -i $BASEDIR/ansible/inventory.yaml -a "pvcreate /dev/$DEVICE"
+  ansible kubenode$1 -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml -a "pvcreate /dev/$DEVICE"
 }
 
 function create_datanode () {
@@ -57,7 +59,7 @@ function create_datanode () {
   ssh-keygen -f "/home/sysadm/.ssh/known_hosts" -R $IPADDR
 
   # Install the Logical Volume Manager (LVM)
-  ansible kubenode$1 -i $BASEDIR/ansible/inventory.yaml -a "apt install -y lvm2 xfsprogs software-properties-common"
+  ansible kubenode$1 -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml -a "apt install -y lvm2 xfsprogs software-properties-common"
 
   # Attach the first data disks
   attach_datadisk $1 vdb 100G
@@ -77,7 +79,7 @@ function create_arbiternode () {
   ssh-keygen -f "/home/sysadm/.ssh/known_hosts" -R $IPADDR
 
   # Install the Logical Volume Manager (LVM)
-  ansible kubenode$1 -i $BASEDIR/ansible/inventory.yaml -a "apt install -y lvm2 xfsprogs software-properties-common"
+  ansible kubenode$1 -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml -a "apt install -y lvm2 xfsprogs software-properties-common"
 
   # Attach the first data disks
   attach_arbiterdisk $1 vdb 20G
@@ -88,10 +90,10 @@ function add_disk () {
   attach_datadisk $1 $2 100G
 
   # Add the new physical volume to the volume group
-  ansible kubenode$1 -i $BASEDIR/ansible/inventory.yaml -a "vgextend vg0 /dev/$2"
+  ansible kubenode$1 -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml -a "vgextend vg0 /dev/$2"
 
   # Add the new physical volume to the logical volume
-  ansible kubenode$1 -i $BASEDIR/ansible/inventory.yaml -a "lvextend /dev/vg0/lv0 /dev/$2 -r"
+  ansible kubenode$1 -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml -a "lvextend /dev/vg0/lv0 /dev/$2 -r"
 }
 
 if [ "$EUID" -ne 0 ]
@@ -102,6 +104,9 @@ fi
 if [ "$1" == "prepare" ]; then
   # Install aptitude which is necessary for Ansible
   apt install aptitude -y
+
+  # Install Python packages
+  pip install pyyaml
 
   # Install the latest version of Ansible from the PPA repository
   apt-add-repository ppa:ansible/ansible
@@ -118,7 +123,7 @@ if [ "$1" == "prepare" ]; then
   chown -R sysadm:sysadm /home/sysadm/.ssh
 
   # Create the LVM
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/host-prepare.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/host-prepare.yaml
 
 elif [ "$1" == "install" ]; then
   # Create the Kubernetes master
@@ -126,8 +131,10 @@ elif [ "$1" == "install" ]; then
   ssh-keygen -f "/root/.ssh/known_hosts" -R $KUBEMASTER_IPADDR
   ssh-keygen -f "/home/sysadm/.ssh/known_hosts" -R $KUBEMASTER_IPADDR
 
-  # Create the arbiter node
-  create_arbiternode 0 110
+  if [ "$STORAGELAYOUT" == "3disk" ]; then
+    # Create the arbiter node
+    create_arbiternode 0 110
+  fi
 
   # Create the first data node
   create_datanode 1 111
@@ -136,40 +143,40 @@ elif [ "$1" == "install" ]; then
   create_datanode 2 112
 
   # Add the nodes to the hosts file of each node
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-hosts.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-hosts.yaml
 
   # Prepare all nodes with a basic install like Docker
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-prepare.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-prepare.yaml
 
   # Install the master node
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-master.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-master.yaml
 
   # Install the worker nodes
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-nodes.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-nodes.yaml
 
   # Install the base components
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-base.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-base.yaml
 
   # Install the GlusterFS Cluster
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-gluster.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-gluster.yaml
 
   # Install Heketi, the storage API for GlusterFS
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-heketi.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-heketi.yaml
 
   # Install the Ingress based on Nginx
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-nginx.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-nginx.yaml
 
   # Install the monitoring solution
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-monitoring.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-monitoring.yaml
 
   # Install Weave
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-weave.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-weave.yaml
 
   # Install the Docker Registry
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-dockerreg.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-dockerreg.yaml
 
   # Deploy custom namespaces
-  ansible-playbook -i $BASEDIR/ansible/inventory.yaml $BASEDIR/ansible/kubernetes-customns.yaml
+  ansible-playbook -i $BASEDIR/ansible/inventory-$STORAGELAYOUT.yaml $BASEDIR/ansible/kubernetes-customns.yaml
 
   reboot
 
@@ -177,8 +184,10 @@ elif [ "$1" == "remove" ]; then
   virsh destroy kubemaster
   virsh undefine kubemaster
 
-  virsh destroy kubenode0
-  virsh undefine kubenode0
+  if [ "$STORAGELAYOUT" == "3disk" ]; then
+    virsh destroy kubenode0
+    virsh undefine kubenode0
+  fi
 
   virsh destroy kubenode1
   virsh undefine kubenode1
@@ -196,29 +205,11 @@ elif [ "$1" == "add-disk" ]; then
   add_disk 1 $2
   add_disk 2 $2
 
-elif [ "$1" == "create-volume" ]; then
-  NS=$2
-  NAME=$3
-
-  ansible kubenodes -i $BASEDIR/ansible/inventory.yaml -m file --args="path=/data/$NS/$NAME state=directory mode=0755"
-  ansible kubenode0 -i $BASEDIR/ansible/inventory.yaml -a "gluster volume create $NS-$NAME replica 2 arbiter 1 kubenode1:/data/$NS/$NAME kubenode2:/data/$NS/$NAME kubenode0:/data/$NS/$NAME --mode=script"
-  ansible kubenode0 -i $BASEDIR/ansible/inventory.yaml -a "gluster volume start $NS-$NAME --mode=script"
-
-elif [ "$1" == "remove-volume" ]; then
-  NS=$2
-  NAME=$3
-
-  ansible kubenode0 -i $BASEDIR/ansible/inventory.yaml -a "gluster volume stop $NS-$NAME --mode=script"
-  ansible kubenode0 -i $BASEDIR/ansible/inventory.yaml -a "gluster volume delete $NS-$NAME --mode=script"
-  ansible kubenodes -i $BASEDIR/ansible/inventory.yaml -m file --args="path=/data/$NS/$NAME state=absent"
-
 else
   echo "Deploys a Kubernetes cluster"
   echo "Usage:"
   echo "  platform.sh prepare"
   echo "  platform.sh install"
   echo "  platform.sh add-disk vd[c-z]"
-  echo "  platform.sh create-volume <namespace> <volname>"
-  echo "  platform.sh remove-volume <namespace> <volname>"
   echo "  platform.sh remove"
 fi
